requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
sqlalchemy==2.0.44
pandas==2.3.3
scikit-learn==1.7.2
joblib==1.5.2
requests==2.32.5
python-dotenv==1.2.1
pydantic-settings==2.1.0
pyyaml==6.0.2
xgboost==2.1.3
lightgbm==4.5.0
pytest==8.3.4
